Renamed to reflect archive lottery functionality.
